[refac] XML parsing structure
