+++ include/sodf/xml/expression_parser.h
@@ -1,0 +1,16 @@
+#ifndef SODF_XML_EXPRESSION_PARSER_H_
+#define SODF_XML_EXPRESSION_PARSER_H_
+
+#include <tinyxml2.h>
+
+namespace sodf {
+namespace xml {
+
+double parseRequiredDoubleExpression(const tinyxml2::XMLElement* elem, const char* attr_name);
+double parseDoubleExpression(const tinyxml2::XMLElement* elem, const char* attr_name, double default_value);
+bool queryDoubleExpression(const tinyxml2::XMLElement* elem, const char* attr_name, double* out_value);
+
+}  // namespace xml
+}  // namespace sodf
+
+#endif  // SODF_XML_EXPRESSION_PARSER_H_
